renamed, rerouted and cleaned up files
renamed all files to be lowercase to match as all other websites have lowercase routes

Changed all routes to better match when we run locally, it will not link to the department server.

Should work on main server as the homepage is matched with the department link, make sure you remove the serverlink from localhost when testing locally

/ route should lead to the startpage not homepage now
