--- Frontend/src/pages/Signup.css
@@ -1,51 +1,0 @@
-/* App.css */
-
-.main {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: calc(100vh - 40px);
-}
-
-.login-container {
-  background-color: #fff;
-  padding: 20px;
-  border-radius: 5px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  text-align: center;
-}
-
-.form-group {
-  margin-bottom: 20px;
-}
-
-.form-group label {
-  display: block;
-  margin-bottom: 5px;
-  text-align: left;
-}
-
-.form-group input {
-  width: 100%;
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-}
-
-.forgot-password {
-  margin-bottom: 10px;
-}
-
-.forgot-password a {
-  color: #333;
-}
-
-.signup-link {
-  margin-top: 20px;
-}
-
-.signup-link a {
-  color: #005BBB;
-  font-weight: bold;
-  text-decoration: none;
-}
