--- Frontend/src/pages/Startcomponents/Banner.js
@@ -1,15 +1,0 @@
-
-import React from 'react';
-import Start from '../Start.css';
-
-function Banner() {
-  return (
-    <section className="banner" >
-      <div className="banner-content">
-      <h2 style={{ fontFamily: 'Crimson Text', fontSize: '48px', color: '#000000' , fontWeight:'400'}}>Know your professor with InSight! <br />  </h2>
-      </div>
-    </section>
-  );
-}
-
-export default Banner;
