--- Frontend/src/pages/Startcomponents/Button.js
@@ -1,12 +1,0 @@
-// components/Button.js
-
-import React from 'react';
-
-function Button({ children }) { 
-  return (
-    <button className="button">{children}</button>
-  );
-}
-
-export default Button;
-
