--- src/webpages/Signup/Signup.jsx
@@ -1,15 +1,0 @@
-// App.js
-
-import React from 'react';
-import './Signup.css';
-import Main from '../../components/Signupcomponents/Main';
-
-function Signup() {
-  return (
-    <div className="App">
-      <Main />
-    </div>
-  );
-}
-
-export default Signup;
